old-tests/webapps/Workers/tests/submissions/Opera/semantics/multiple-workers/002.html

<!--
postMessage(1);
/*
-->
<!doctype html>
<title>creating 3 sibling dedicated workers</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function() {
  var w1 = new Worker('#');
  var w2 = new Worker('#');
  var w3 = new Worker('#');
  var gotW1 = false;
  var gotW2 = false;
  var gotW3 = false;
  w1.onmessage = function(e) { gotW1 = true; };
  w2.onmessage = function(e) { gotW2 = true; };
  w3.onmessage = function(e) { gotW3 = true; };
  setTimeout(this.step_func(function() {
    assert_true(gotW1, '1');
    assert_true(gotW2, '2');
    assert_true(gotW3, '3');
    this.done();
  }), 300);
});
</script>
<!--
*/
//-->
